+++ src/App.js
@@ -1,5 +1,4 @@
 import React, { Component } from 'react';
-import logo from './logo.svg';
 import './App.css';
 import DogsListContainer from './components/DogsListContainer';
