+++ src/components/DogsListContainer.js
@@ -1,28 +1,33 @@
 import React, { Component } from 'react'
-import request from 'superagent'
 import DogsList from './DogsList'
+import { getDogsList } from '../actions/DogsListAction'
 
 export default class DogsListContainer extends Component {
-    state = { dogBreeds: null }
+    // state = { dogBreeds: null }
 
     componentDidMount() {
-        request
-            .get('https://dog.ceo/api/breeds/list/all')
-            .then(response => {
-                const breeds = Object.keys(response.body.message)
-                this.updateBreeds(breeds)
-            })
-            .catch(console.error)
+        this.props.getDogsList()
+        // request
+        //     .get('https://dog.ceo/api/breeds/list/all')
+        //     .then(response => {
+        //         const breeds = Object.keys(response.body.message)
+        //         this.updateBreeds(breeds)
+        //     })
+        //     .catch(console.error)
     }
 
-    updateBreeds(breeds) {
-        this.setState({
-            dogBreeds: breeds
-        })
-    }
+    // updateBreeds(breeds) {
+    //     this.setState({
+    //         dogBreeds: breeds
+    //     })
+    // }
 
     render() {
-        return <DogsList dogBreeds={this.state.dogBreeds} />
-
+        const { dogsList } = this.props
+        return (
+            { dogsList }
+            // <DogsList />)
+            // return <DogsList dogBreeds={this.state.dogBreeds} />
+        )
     }
 }
