+++ src/reducers/index.js
@@ -1,0 +1,6 @@
+import DogsListReducer from './DogsListReducer'
+import { combineReducers } from 'redux';
+
+export default combineReducers({
+    DogsListReducer
+})
