--- src/reducers/reducer.js
@@ -1,8 +1,0 @@
-const reducer = (state = [], action = {}) => {
-  switch (action.type) {
-  default:
-    return state
-  }
-}
-
-export default reducer
